refactor: simplify processor settings (#122)
- replace LLM and diarization config wrappers with direct processor objects
- move provider/model construction to the CLI for requested optional features
- fold tiny helper modules into their owning modules and delete unused indirection

+++ webinar_transcriber/tests/test_video_pipeline.py
@@ -142,7 +142,7 @@ class TestDetectScenes:
     @pytest.mark.slow
     def test_detect_scenes_finds_synthetic_alternating_color_scene(self, tmp_path: Path) -> None:
         video_path = tmp_path / "alternating-color.mp4"
-        progress_updates: list[tuple[int, int]] = []
+        progress_updates: list[tuple[float, int]] = []
         _write_synthetic_video(video_path, [0, 255] * 5)
 
         scenes = detect_scenes(
